pyproject.toml
[project]
name = "openweb-ui-classification-examples"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.2.0",
    "langchain-chroma>=0.1.0",
    "langchain-huggingface>=0.0.3",
    "langchain-openai>=0.1.0",
    "litellm>=1.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "sentence-transformers>=2.2.0",
]
